added auth

+++ graphql/resolvers/deleteSpot.js
@@ -1,27 +1,28 @@
-const Spot = require("../../models/spot")
+const Spot = require('../../models/spot');
 
+module.exports = async function deleteSpot({ _id }, req, res) {
+  if (!res.request.isAuth) {
+    const error = new Error('Not authenticated!');
+    error.code = 401;
+    throw error;
+  }
 
-module.exports = async function deleteSpot({ _id }, req) {
+  const user_id = '5dd741edb65656757bb522c6';
+  let spot = '';
 
-  const user_id = '5dd741edb65656757bb522c6'
-  let spot = ''
-
-  try{
+  try {
     spot = await Spot.findById(_id);
-  }catch(e){
-    throw new Error('Unable to find spot')
+  } catch (e) {
+    throw new Error('Unable to find spot');
   }
 
-  try{
+  try {
     if (spot.owner.toString() === user_id.toString()) {
-      await Spot.findByIdAndDelete({_id});
+      await Spot.findByIdAndDelete({ _id });
     }
-  }catch(e){
-    throw new Error('You do not own this spot')
+  } catch (e) {
+    throw new Error('You do not own this spot');
   }
-  
-  
-  return "Spot was deleted";
-  
-      
-}
+
+  return 'Spot was deleted';
+};
