+++ graphql/resolvers/deleteUser.js
@@ -1,17 +1,21 @@
-const User = require("../../models/User")
+const User = require('../../models/User');
 
-module.exports = async function deleteUser({ _id }, req) {
-    
-    try{
-        const user = await User.findById(_id);
-    }
-    catch (e){
-        throw new Error ('Could not find user')
-    }
+module.exports = async function deleteUser({ _id }, req, res) {
+  if (!res.request.isAuth) {
+    const error = new Error('Not authenticated!');
+    error.code = 401;
+    throw error;
+  }
 
-    await User.findByIdAndDelete({
-        _id
-    });
+  try {
+    const user = await User.findById(_id);
+  } catch (e) {
+    throw new Error('Could not find user');
+  }
 
-    return "User was deleted";
-}
+  await User.findByIdAndDelete({
+    _id,
+  });
+
+  return 'User was deleted';
+};
